moved EquipmentModification to new folder
